[fixes #50392275] - show more helpful error message when dictionary words file is not installed
Also retry to generate shared secrets if file exists but is empty

+++ lib/bootstrap-cf-plugin/shared_secrets_file.rb
@@ -6,11 +6,14 @@ module BootstrapCfPlugin
     def self.random_string
       Haddock::Password.delimiters = %q#`~!@$%^&*()-_=+[{]}\;:'",<.>/?#
       Haddock::Password.generate(PASSWORD_LENGTH)
+    rescue Haddock::Password::NoWordsError
+      puts("We can't find your dictionary words file.   Please make sure you have one installed... this is usually part of the wamerican pacakge on your system")
+      exit(1)
     end
 
     def self.find_or_create(filename)
       template_path = File.join(File.dirname(__FILE__), "..", "..", "templates", "cf-shared-secrets.yml.erb")
-      unless File.exists?(filename)
+      unless File.exists?(filename) && !File.read(filename).empty?
         File.open(filename, "w") do |f|
           f.write ERB.new(File.read(template_path)).result(binding)
         end
